api/requirements.txt
fastapi[all]
uvicorn[standard]==0.17.6
psycopg[binary,pool]==3.1.17
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
